requirements.txt
babel==2.8.0
captum==0.7.0
cloudpickle==1.6.0
deprecation==2.1.0
filelock==3.0.12
h5py==3.9.0
htcondor==10.7.0
imbalanced-learn==0.7.0
Jinja2==2.11.2
joblib==1.3.2
MarkupSafe==1.1.1
matplotlib==3.7.2
nibabel==3.2.1
nilearn==0.10.1
numpy==1.26.4
packaging==23.1
pandas==1.2.0
posix-ipc==1.0.5
pytest==8.1.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.5
scikit-learn==1.3.0
scipy==1.11.2
signac==1.5.1
signac-flow==0.11.0
six==1.15.0
sphinx==3.4.3
threadpoolctl==2.1.0
torch==1.12.1
tqdm==4.55.0
typing-extensions==3.7.4.3